+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 
 ### Added
 - Week Number action to Date Time tool.
+- Fluent theme, other windows (settings, management, etc) and menus now match users Windows theme.
 - 
 
 ### Changed